throw error if user and domain are empty, factor out processor code to OneOnOneResolver, add UTs - WPB-10191 (#1994)
