+++ WireDomain/Tests/WireDomainTests/Event Processing/UserEventProcessor/UserConnectionEventProcessorTests.swift
@@ -16,82 +16,48 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import Foundation
 import WireAPI
-import WireAPISupport
-import WireDataModel
-import WireDataModelSupport
+@testable import WireDomain
 import WireDomainSupport
 import XCTest
-
-@testable import WireDomain
 
 final class UserConnectionEventProcessorTests: XCTestCase {
     var sut: UserConnectionEventProcessor!
 
-    var coreDataStack: CoreDataStack!
-    var coreDataStackHelper: CoreDataStackHelper!
-    var modelHelper: ModelHelper!
     var connectionsRepository: MockConnectionsRepositoryProtocol!
-    var userRepository: MockUserRepositoryProtocol!
-    var conversationsRepository: MockConversationRepositoryProtocol!
-    var mlsService: MockMLSServiceInterface!
-
-    var context: NSManagedObjectContext {
-        coreDataStack.syncContext
-    }
+    var oneOnOneResolver: MockOneOnOneResolverProtocol!
 
     override func setUp() async throws {
         try await super.setUp()
-        coreDataStackHelper = CoreDataStackHelper()
-        modelHelper = ModelHelper()
-        coreDataStack = try await coreDataStackHelper.createStack()
         connectionsRepository = MockConnectionsRepositoryProtocol()
-        userRepository = MockUserRepositoryProtocol()
-        conversationsRepository = MockConversationRepositoryProtocol()
-        mlsService = MockMLSServiceInterface()
+        oneOnOneResolver = MockOneOnOneResolverProtocol()
         sut = UserConnectionEventProcessor(
-            context: context,
             connectionsRepository: connectionsRepository,
-            userRepository: userRepository,
-            conversationsRepository: conversationsRepository,
-            mlsService: mlsService
+            oneOnOneResolver: oneOnOneResolver
         )
-
-        DeveloperFlag.storage = UserDefaults(suiteName: Scaffolding.defaultsSuiteName)!
-        var flag = DeveloperFlag.enableMLSSupport
-        flag.isOn = true
     }
 
     override func tearDown() async throws {
         try await super.tearDown()
-        coreDataStack = nil
-        sut = nil
-        modelHelper = nil
-        try coreDataStackHelper.cleanupDirectory()
-        DeveloperFlag.storage.removePersistentDomain(forName: Scaffolding.defaultsSuiteName)
-        coreDataStackHelper = nil
         connectionsRepository = nil
-        userRepository = nil
-        conversationsRepository = nil
-        mlsService = nil
+        oneOnOneResolver = nil
+        sut = nil
     }
 
     // MARK: - Tests
 
-    func testProcessEvent_It_Invokes_Update_Connection_Repo_Method() async throws {
+    func testProcessEvent_It_Invokes_Methods() async throws {
         // Given
-
-        let connectionStatus = ConnectionStatus.pending
 
         let event = UserConnectionEvent(
             userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
+            connection: Scaffolding.connection
         )
 
         // Mock
 
         connectionsRepository.updateConnection_MockMethod = { _ in }
+        oneOnOneResolver.invoke_MockMethod = {}
 
         // When
 
@@ -100,354 +66,26 @@ final class UserConnectionEventProcessorTests: XCTestCase {
         // Then
 
         XCTAssertEqual(connectionsRepository.updateConnection_Invocations, [event.connection])
-    }
-
-    func testProcessEvent_It_Does_Not_Migrate_MLS_Conversation() async throws {
-        // Given
-
-        let connectionStatus = ConnectionStatus.accepted
-
-        let event = UserConnectionEvent(
-            userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
-        )
-
-        let commonProtocol = WireDataModel.MessageProtocol.mls
-
-        let (selfUser, user, mlsOneOnOneConversation) = try setupManagedObjects(
-            selfUserProtocol: commonProtocol,
-            userProtocol: commonProtocol
-        )
-
-        let mlsConversationExists = true /// should not migrate MLS conversation
-
-        // Mock
-
-        setupMock(
-            selfUser: selfUser,
-            user: user,
-            mlsOneOnOneConversation: mlsOneOnOneConversation,
-            mlsConversationExists: mlsConversationExists
-        )
-
-        // When
-
-        try await sut.processEvent(event)
-
-        // Then
-
-        XCTAssert(mlsService.establishGroupForWith_Invocations.isEmpty)
-        XCTAssert(mlsService.joinGroupWith_Invocations.isEmpty)
-    }
-
-    func testProcessEvent_It_Resolves_MLS_Conversation_Epoch_Zero() async throws {
-        // Given
-
-        let connectionStatus = ConnectionStatus.accepted
-
-        let event = UserConnectionEvent(
-            userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
-        )
-
-        let commonProtocol = WireDataModel.MessageProtocol.mls
-        let mlsEpoch: UInt64 = 0
-
-        let (selfUser, user, mlsOneOnOneConversation) = try setupManagedObjects(
-            selfUserProtocol: commonProtocol,
-            userProtocol: commonProtocol,
-            mlsEpoch: mlsEpoch
-        )
-
-        // Mock
-
-        setupMock(
-            selfUser: selfUser,
-            user: user,
-            mlsOneOnOneConversation: mlsOneOnOneConversation
-        )
-
-        // When
-
-        try await sut.processEvent(event)
-
-        // Then
-
-        XCTAssertEqual(mlsService.establishGroupForWith_Invocations.count, 1)
-        let createGroupInvocation = try XCTUnwrap(
-            mlsService.establishGroupForWith_Invocations.first
-        )
-
-        XCTAssertEqual(createGroupInvocation.groupID, Scaffolding.mlsGroupID)
-        XCTAssertEqual(
-            createGroupInvocation.users,
-            [MLSUser(Scaffolding.receiverQualifiedID.toDomainModel())]
-        )
-        XCTAssertEqual(mlsOneOnOneConversation.ciphersuite, Scaffolding.ciphersuite)
-        XCTAssertEqual(mlsOneOnOneConversation.mlsStatus, .ready)
-        XCTAssertEqual(mlsOneOnOneConversation.isForcedReadOnly, false)
-        XCTAssertEqual(mlsOneOnOneConversation.needsToBeUpdatedFromBackend, true)
-        XCTAssertEqual(user.oneOnOneConversation, mlsOneOnOneConversation)
-        XCTAssertEqual(mlsOneOnOneConversation.oneOnOneUser, user)
-    }
-
-    func testProcessEvent_It_Resolves_MLS_Conversation_Epoch_Not_Zero() async throws {
-        // Given
-
-        let connectionStatus = ConnectionStatus.accepted
-
-        let event = UserConnectionEvent(
-            userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
-        )
-
-        let commonProtocol = WireDataModel.MessageProtocol.mls
-        let mlsEpoch: UInt64 = 1
-
-        let (selfUser, user, mlsOneOnOneConversation) = try setupManagedObjects(
-            selfUserProtocol: commonProtocol,
-            userProtocol: commonProtocol,
-            mlsEpoch: mlsEpoch
-        )
-
-        // Mock
-
-        setupMock(
-            selfUser: selfUser,
-            user: user,
-            mlsOneOnOneConversation: mlsOneOnOneConversation
-        )
-
-        // When
-
-        try await sut.processEvent(event)
-
-        // Then
-
-        XCTAssertEqual(mlsService.joinGroupWith_Invocations.count, 1)
-        let invokedMLSGroupID = try XCTUnwrap(mlsService.joinGroupWith_Invocations.first)
-        XCTAssertEqual(invokedMLSGroupID, Scaffolding.mlsGroupID)
-        XCTAssertEqual(user.oneOnOneConversation, mlsOneOnOneConversation)
-        XCTAssertEqual(mlsOneOnOneConversation.oneOnOneUser, user)
-    }
-
-    func testProcessEvent_It_Migrates_Proteus_Messages_To_MLS_Conversation() async throws {
-        // Given
-
-        let connectionStatus = ConnectionStatus.accepted
-
-        let event = UserConnectionEvent(
-            userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
-        )
-
-        let commonProtocol = WireDataModel.MessageProtocol.mls
-
-        let (selfUser, user, mlsOneOnOneConversation) = try setupManagedObjects(
-            selfUserProtocol: commonProtocol,
-            userProtocol: commonProtocol
-        )
-
-        // Mock
-
-        setupMock(
-            selfUser: selfUser,
-            user: user,
-            mlsOneOnOneConversation: mlsOneOnOneConversation
-        )
-
-        // When
-
-        try await sut.processEvent(event)
-
-        // Then
-
-        let migratedMessagesTexts = mlsOneOnOneConversation.allMessages
-            .compactMap(\.textMessageData)
-            .compactMap(\.messageText)
-            .sorted()
-
-        /// Ensuring proteus messages were migrated to MLS conversation.
-        XCTAssertEqual(migratedMessagesTexts.first, "Hello")
-        XCTAssertEqual(migratedMessagesTexts.last, "World!")
-    }
-
-    func testProcessEvent_It_Resolves_Proteus_Conversation() async throws {
-        // Given
-
-        let connectionStatus = ConnectionStatus.accepted
-
-        let event = UserConnectionEvent(
-            userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
-        )
-
-        let commonProtocol = WireDataModel.MessageProtocol.proteus
-
-        let (selfUser, user, mlsOneOnOneConversation) = try setupManagedObjects(
-            selfUserProtocol: commonProtocol,
-            userProtocol: commonProtocol
-        )
-
-        XCTAssertEqual(user.oneOnOneConversation?.isForcedReadOnly, true)
-
-        // Mock
-
-        setupMock(
-            selfUser: selfUser,
-            user: user,
-            mlsOneOnOneConversation: mlsOneOnOneConversation
-        )
-
-        // When
-
-        try await sut.processEvent(event)
-
-        // Then
-
-        XCTAssertEqual(user.oneOnOneConversation?.isForcedReadOnly, false)
-    }
-
-    func testProcessEvent_It_Resolves_Conversation_With_No_Common_Protocol() async throws {
-        // Given
-
-        let connectionStatus = ConnectionStatus.accepted
-
-        let event = UserConnectionEvent(
-            userName: Scaffolding.username,
-            connection: setupConnection(status: connectionStatus)
-        )
-
-        let forcedReadOnly = false
-
-        let (selfUser, user, mlsOneOnOneConversation) = try setupManagedObjects(
-            selfUserProtocol: .mls,
-            userProtocol: .proteus,
-            forcedReadOnly: forcedReadOnly
-        )
-
-        XCTAssertEqual(user.oneOnOneConversation?.isForcedReadOnly, false)
-
-        // Mock
-
-        setupMock(
-            selfUser: selfUser,
-            user: user,
-            mlsOneOnOneConversation: mlsOneOnOneConversation
-        )
-
-        // When
-
-        try await sut.processEvent(event)
-
-        // Then
-
-        let lastMessage = try XCTUnwrap(user.oneOnOneConversation?.lastMessage as? ZMSystemMessage)
-        XCTAssertEqual(lastMessage.systemMessageType, .mlsNotSupportedOtherUser)
-        XCTAssertEqual(user.oneOnOneConversation?.isForcedReadOnly, true)
-    }
-
-    // MARK: - Setup
-
-    private func setupManagedObjects(
-        selfUserProtocol: WireDataModel.MessageProtocol,
-        userProtocol: WireDataModel.MessageProtocol,
-        forcedReadOnly: Bool = true,
-        mlsEpoch: UInt64 = 0
-    ) throws -> (selfUser: ZMUser,
-                 user: ZMUser,
-                 mlsConversation: ZMConversation) {
-        let user = modelHelper.createUser(
-            id: Scaffolding.receiverQualifiedID.uuid,
-            domain: Scaffolding.receiverQualifiedID.domain,
-            in: context
-        )
-
-        user.supportedProtocols = [userProtocol]
-
-        let selfUser = modelHelper.createSelfUser(
-            id: UUID(),
-            domain: nil,
-            in: context
-        )
-
-        selfUser.supportedProtocols = [selfUserProtocol]
-
-        let proteusConversation = modelHelper.createOneOnOne(
-            with: selfUser,
-            in: context
-        )
-
-        proteusConversation.isForcedReadOnly = forcedReadOnly
-        user.oneOnOneConversation = proteusConversation
-
-        try proteusConversation.appendText(content: "Hello")
-        try proteusConversation.appendText(content: "World!")
-
-        let mlsOneOnOneConversation = modelHelper.createMLSConversation(
-            mlsGroupID: Scaffolding.mlsGroupID,
-            mlsStatus: .pendingJoin,
-            conversationType: .oneOnOne,
-            epoch: mlsEpoch,
-            in: context
-        )
-
-        return (selfUser, user, mlsOneOnOneConversation)
-    }
-
-    private func setupMock(
-        selfUser: ZMUser,
-        user: ZMUser,
-        mlsOneOnOneConversation: ZMConversation,
-        mlsConversationExists: Bool = false
-    ) {
-        userRepository.fetchUserWithDomain_MockValue = user
-        userRepository.fetchSelfUser_MockValue = selfUser
-
-        connectionsRepository.updateConnection_MockMethod = { _ in }
-        conversationsRepository.pullMLSOneToOneConversationUserIDDomain_MockValue = Scaffolding.conversationID.uuidString
-        conversationsRepository.fetchMLSConversationWith_MockValue = mlsOneOnOneConversation
-
-        mlsService.establishGroupForWith_MockValue = Scaffolding.ciphersuite
-        mlsService.conversationExistsGroupID_MockValue = mlsConversationExists
-        mlsService.joinGroupWith_MockMethod = { _ in }
-    }
-
-    private func setupConnection(status: ConnectionStatus) -> Connection {
-        Connection(
-            senderID: Scaffolding.senderID,
-            receiverID: Scaffolding.receiverID,
-            receiverQualifiedID: Scaffolding.receiverQualifiedID,
-            conversationID: Scaffolding.conversationID,
-            qualifiedConversationID: Scaffolding.qualifiedConversationID,
-            lastUpdate: .now,
-            status: status
-        )
+        XCTAssertEqual(oneOnOneResolver.invoke_Invocations.count, 1)
     }
 
     private enum Scaffolding {
         static let username = "username"
-        static let senderID = UUID()
-        static let receiverID = UUID()
-        static let receiverQualifiedID = WireAPI.QualifiedID(
-            uuid: receiverID,
-            domain: "domain.com"
+        static let connection = Connection(
+            senderID: UUID(),
+            receiverID: UUID(),
+            receiverQualifiedID: WireAPI.QualifiedID(
+                uuid: UUID(),
+                domain: "domain.com"
+            ),
+            conversationID: UUID(),
+            qualifiedConversationID: WireAPI.QualifiedID(
+                uuid: UUID(),
+                domain: "domain.com"
+            ),
+            lastUpdate: .now,
+            status: .accepted
         )
-        static let conversationID = UUID()
-        static let qualifiedConversationID = WireAPI.QualifiedID(
-            uuid: conversationID,
-            domain: "domain.com"
-        )
-
-        static let base64EncodedString = "pQABARn//wKhAFggHsa0CszLXYLFcOzg8AA//E1+Dl1rDHQ5iuk44X0/PNYDoQChAFgg309rkhG6SglemG6kWae81P1HtQPx9lyb6wExTovhU4cE9g=="
-
-        static let ciphersuite = WireDataModel.MLSCipherSuite.MLS_256_DHKEMP521_AES256GCM_SHA512_P521
-
-        static let mlsGroupID = WireDataModel.MLSGroupID(
-            base64Encoded: base64EncodedString
-        )!
-
-        static let defaultsSuiteName = UUID().uuidString
     }
 
 }
